Add end-to-end example compiling through IREE.
An example of a single brgemm op is added that runs all the way through IREE up until lowering to the libxsmm function calls.
This will compose out of the box with IREE's parallelism (note: undesirable alloc needs to be fixed).

Connecting these pieces required registering linalgx, tpp, xsmm in LLVMCPULowerExecutableTarget.cpp to allow such ops to be produced by the transform dialect.

2 additional things are needed to get that executing properly:

  1. Lower to library directly using the extract_strided_metadata rather than relying on `memref<*xf32>`.
  2. Load libxsmm in IREE and run through it.