WIP - Add configurable packing/padding example that connects all the way to codegen
This revision adds an mlir file to accelerate bringing up the cuda codegen path for a
packed matmul. The packing is configurable, instructions are provided in ample comments in the mlir file.

Repros for IR dumps and generating PTX (subject to the 2 temporary IREE caveats mentioned in #12076).

Preprocessing with generalized packing.
```
  iree-opt tests/transform_dialect/cuda/matmul-packing.mlir --iree-transform-dialect-interpreter --transform-dialect-drop-schedule
```

Dump the IR after dispatch region formation and focus our attention on the shape of the generic.
```
  iree-opt tests/transform_dialect/cpu/contraction-packing.mlir --iree-transform-dialect-interpreter --transform-dialect-drop-schedule | \
  iree-opt  --iree-hal-target-backends=cuda --iree-abi-transformation-pipeline  --iree-flow-transformation-pipeline  --iree-stream-transformation-pipeline  --iree-hal-configuration-pipeline | \
  grep -5 linalg\.generic
```

Compile e2e using the default cuda backend path: this is subject to the 2 temporary IREE caveats mentioned in #12076:
  - the flow pipeline calls InterchangeGenericOps: this is not a deal breaker but changes the iteration order of the packed op a bit
  - difficulties to lower either tensor.pack/unpack or linalg_ext.pack/unpack forms force us to lower to linalg.fill/linalg.transpose/etc
```
  iree-opt tests/transform_dialect/cuda/matmul-packing.mlir --iree-transform-dialect-interpreter --transform-dialect-drop-schedule | \
  iree-compile -  --iree-hal-target-backends=cuda
```